refactor: 컴포넌트를 barrel 방식으로 수정하기
