+++ app/my-profile/activity.tsx
@@ -3,8 +3,7 @@ import React from 'react';
 import { View } from 'react-native';
 import { ScrollView } from "react-native-gesture-handler";
 
-import ImageButton from '@/components/my-profile/activity/image-button';
-import MenuGroupList from '@/components/my-profile/activity/menu-group-list';
+import { ImageButton, MenuGroupList } from '@/components/my-profile/activity';
 
 // TODO: 가상 데이터를 실제 데이터로 교체하기
 const profileMockData = { id: 1, name: '돌돌이' };
