+++ components/my-profile/activity/index.ts
@@ -1,0 +1,10 @@
+import Header from "./header";
+import ImageButton from "./image-button";
+import MenuGroupList from "./menu-group-list";
+
+
+export {
+  Header,
+  ImageButton,
+  MenuGroupList,
+}
